Tasks complete with autoFeatures.
